pip-requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
dnspython==1.16.0
eventlet==0.25.0
future==0.17.1
get==2019.4.13
greenlet==0.4.15
idna==2.8
monotonic==1.5
post==2019.4.13
public==2019.4.13
pycparser==2.19
pyTelegramBotAPI==3.6.6
python-telegram-bot==11.1.0
query-string==2019.4.13
request==2019.4.13
requests==2.7.0
six==1.12.0
SQLAlchemy==1.3.7
telebot==0.0.3
urllib3==1.25.3
